requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
billiard==3.5.0.5
boto3==1.9.109
botocore==1.12.109
celery==4.2.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.6.1
Django==2.1.7
django-crispy-forms==1.7.2
django-filter==2.1.0
django-registration-redux==2.5
django-rest-framework==0.1.0
django-rest-knox==3.6.0
django-s3-storage==0.12.4
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
docutils==0.14
drf-multiple-settings==1.0.0
drf-yasg==1.13.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
kombu==4.4.0
MarkupSafe==1.1.1
Pillow==5.4.1
psycopg2-binary==2.7.7
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
ruamel.yaml==0.15.89
s3transfer==0.2.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
